Pin scikit-build==0.11.0 in Appveyor
As of version 0.12.0 scikit-build strips of the Win64 suffix in the
generator name [1], which makes scikit-build pick up the 32bit version
on 64bit builds. Because the core was correctly built as x64, we get all
sorts of linking errors when linking the python extension, naturally.

There is probably a correct way of dealing with this, but for now
pinning the version should be fine.

[1] https://github.com/scikit-build/scikit-build/blob/3e3dc51dc04aafd0bb694139bbe5d2ebf0cac466/skbuild/platform_specifics/abstract.py#L292