Add matlab job
There is no plan to do anything with matlab bindings (we are not sure if
we won't drop them whatsoever), but it would be useful to know if we
accidentally break something in the future.

Github actions allow us to use matlab.
While originally intended way of running tests with ctests doesn't work
for some reason (probably due to matlab setup in github actions), we
still can run scripts there directly as commands.