Add benchmark tests
Should allow us to make sure future development is not causing major
drop in calculation speed.

Currently only most straightforward cases are tested. Threshold is set
to 25% of allowed slowness, which should be around x2 observed worst
differences between two runs on the same sha.

File sizes are reduced significantly to 1) fit into github actions
runner 2) to assure stable enough environment in order to reduce speed
fluctuations.

Currently actions is also set to run on pull requests. We will see how
often that fails and adjust triggers accordingly.

Tests are considered to be secondary to main pytest suite and won't be
run by default. Thus pytest-benchmark dependency is not considered to be
a requirement for build and is not mentioned in the Readme.