Add bigendian workflow job
We build and test wheels with cibuildwheel, but as no emulation happens
there, tests are not actually run on bigendian system.

In this new workflow we use qemu to emulate the setup. Run time is long,
but we want this workflow anyway as our intention is to support
bigendian environments.

The most time-consuming operation is building numpy for s390x. Thus we
run job only on merge to main as job is not expected to be failing often
anyway. Alternatively we could build and cache numpy in github actions,
but we decided to go with simpler solution for the moment.

(cherry picked from commit cd874f4581d61c456159bea3dcbcbb2c53591d60)