Extract segy_set_endianness from segy_set_format
Making format function to be responsible for endianness is confusing.

- While in C code SEGY_LSB is defined as 0x01 0x00 and SEGY_MSB is
defined as 0x02 0x00, in python little is set to 0x01 0x00 while big is
set to 0x00 0x00.
- SEGY_MSB path has never been used in the runtime.
- Corresponding function segy_format retrieves only value of the format
itself, not value on endianness.

Hence responsibility is split to different functions and identificators
are unified.

Standard defines 3 possible endianness cases, thus int is still used
instead of a boolean, in case third one must be implemented. It must be
noted however that in current code lsb flag, even though is int, is used
as bool.

Change affects python bindings, which are fixed.
There seems to be no wrapper around that in experimental c++ bindings,
so both format and endianness are left as defaults.