Correct multiformat int24 tests
According to documentation, int24 test type is "quite incomplete".
(See more information on original file creation at
6fe177c48d114f8a038c65381a5ed7ae1b0d31bd)
Unfortunately that leads to confusion when trying to comprehend test
behavior. This commit attempts to alleviate some of the confusion.

1.
Intention of multiformat tests is to assure that files in all formats
are read correctly. To test it f3.sgy file (format 3) was converted to
all other formats.
In theory we would like to assure that each sample value x in format n
equals corresponding sample value in format 3.
Instead we verify not that values are the same, but just that some bytes
are.

(Examples below are laid out as bigendian).

1a:
From 2-byte signed integer f3.sgy file, 3-byte signed integer files
were created by setting the most significant byte to 0.
While numbers 0x12 0x34 (4660 dec) and 0x00 0x12 0x34 (4660 dec) are
equivalent when read as signed 2-byte 2's complement int and signed
3-byte 2's complement int,
0xAB 0xCD (-21555 dec) and 0x00 0xAB 0xCD (43981 dec) are not.

In the code casting from f3 format (int16) to int24 is implemented such
that most significant byte is always set to 0. Thus f3 value, regardless
of sign, is transformed to the same value that was written into format7
file.

Behavior was intended, but is confusing, especially because 4-byte and
8-byte signed/unsigned int files are sign-aware. To improve this, int24
must be made sign-aware and format 7/15 files must be fixed by setting
0xFF instead of 0x00 byte for negative numbers - turn 0x00 0xAB 0xCD
(43981 dec) to 0xFF 0xAB 0xCD (-21555 dec).

This is fixed with a one-time script equivalent of

for i in range(samples_value):
    bytes_read = f.read(3)
    if bytes_read[1] & 0x80:
        f.seek(-3, os.SEEK_CUR)
        if bigendian:
            f.write(b'\xFF' + bytes_read[1:])
        else:
            f.write(bytes_read[:2] + b'\xFF')

1b:
Similar value inequality happens for 1-byte formats 8 and 16 also.
2-byte integer file was converted to 1-byte integer files by cutting of
the most significant byte. Usually values in f3.sgy do not have most
significant byte = 0, so 2-byte int value 0x12 0x34 in f3 became 0x34.
0x12 0x34 is not equal 0x34, so values in the files are not equal.
However comparison is done by casting f3 values into 1-byte type value,
which by default will cut off the most significant byte.
So in this case only the least significant bytes are compared and thus
test passes.

1c:
As original file contains signed integers, they are impossible to
correctly represent as unsigned integers. Thus rules are defined by
conversion of signed type to unsigned. So for example signed 2-byte int
0xF5 0xCE (-2610) and unsigned 4-byte int 0xFF 0xFF 0xF5 0xCE
(4294964686) are considered equal.

A cleaner approach would require creation of valid 1-byte files and
their conversion to different formats, but it would require effort,
which currently is not warranted as existing solution is good enough.

2. conversion to int32 is only used during comparison, so it could be
implemented any way that can state whether internal bytes are the same
or not. So while not required by tests, implementation is changed to
actually represent the number that was provided.