Build wheel in benchmark CI
Current logic is flawed.
Intention of benchmark tests was to build library from two different
commits. However only C/C++ library was built separately, python code
was not touched and included.

When current commit's directory was checked out to run the fresh tests,
so was other python code: tests were always run on the same python part
of the library.

To fix that, build a local wheel from existing code and make sure it is
being used instead of local code.