Add tests including public headers
These tests try to check whether the inclusion of segyio headers that
are part of segyio's public interface prevent other projects from
compiling. We had cases in which our public facing C and C++ headers did
not define all required includes. This could lead to failures depending
on the compiler and what other includes were and in which order when a
segyio header was included.

Testing for the correct headers to be included appears to be non-trivial
and it is not 100% clear if this test is sufficiently accurate to
prevent further problems in the future. At least the test tries to
approximate how segyio's C/C++ interfaces are included in other projects
such as the the si4ti project [1] that uses the C++ interface. If our
tests is too simple, we could also try a tool like "include-what-you-use"
[2]. If our test does not provide sufficient value, we can remove it
again in the future.

Note, the added test should fail at compile time if anything is wrong
with the includes.

[1]: https://github.com/equinor/si4ti
[2]: https://github.com/include-what-you-use/include-what-you-use