Test large file through python
Current test that is supposed to read the large file actually doesn't do
it. All it does is testing how well segy_seek and segy_ftell functions
match each other.

Giving that segy_ftell is not used in any other place in the codebase
other than in this one test function, usefulness of this test is
questionable.

Hence test was substituted with actual large file test run through
python.

Further plan is to add more tests to this file, for example to test
correct processing when max samples number would exceed 16-bit int.