Pin cbuildwheel version in benchmark tests
New release of cbuildwheel updated cmake version to 4.0.0 (probably
because manylinix did).
Cmake 4.0 drops support for cmake < 3.5.

In our old "reference" code, which we would prefer to test against until
segyio 2.0 is released, one of the cmake files still declares minimum
required version to be 2.x and therefore is affected.

The easiest workaround is to pin cbuildwheel to latest cmake-compatible
version v2.23.2.