Change defaults for binary fields on file creation
ntrpr and nart:
Currently number of traces is stored into those fields on file creation.

Not only this is wrong from the domain perspective, it is also wrong
from the data integrity point of view. Both fields have space just for 2
bytes, while number of traces can be 4-8 bytes. In this case values
getting stored there are truncated and become meaningless.

As both fields are mandatory for prestack data, default value of 1 is
set for ntrpr and nart is left to be 0.
Value can always be overwritten by the user.

hns and nso:
Are not set if value cannot be stored on 2 bytes.

fold:
Is expected to be equal to ntrpr.

tsort:
Field is used incorrectly. It does not describe the way segy file is
sorted, but describes sorting within ensemble. Default value should be
"horizontally stacked".