Use fseeko and _fseeki64 for file seek
Currently we use segy_seek on with SEEK_SET whence, but once this logic
is used in file datasource seek, we would need to apply it at least once
with SEEK_CUR too.
That would mean that for systems where long is 32-bit we would have to
implement logic for seeking to positive and negative long long position
with regards to SEEK_CUR.

It seems a bit weird that we have to reinvent the wheel when there exist
_fseeki64 function on Windows and fseeko for Linux.

It is unclear why it was not used originally. _ftelli64 was used only in
"test" function with justification that it is "a little rough".

Currently those functions are used widely in the community, so we can
assume them generally safe.

fseek is exchanged with fseeko. It will now handle 32-bit Linux cases
instead of custom logic.