Add python segy_datasource implementation
Functions are tested with various datasources, but still may contain
surprises as semantics of Python-C bindings is a bit complex.
Switching to pybind11 in the future could simplify the flow.

Note that read copies the buffer using memcpy. Attempt to use readinto
when available resulted in worse performance (for unknown reason) and
thus was abandoned.