Make segyfd work both on streams and files
segyfd is reused both for file and stream input datasources as creating
separate modules seems unnecessary.

All arguments in segyiofd init are turned from positional to keywords.
File and mode or stream are expected to represent input datasource.
Endianness is already set to be an optional argument as well
- segy 2.1 opens to discover endianness without any need for user input.

This change does not affect users of python interface, only direct users
of C-Python bindings.