Add memory datasource
Some files might fit into memory just fine. Users could use standard
stream datasource with BufferIO to deal with such files. However by
directly sharing python memory with C code via PyBuffer object, we would
spare C-Python back and forth trips, which should benefit performance.

This option is available only for file reading and updating (which does
not change size of the data and thus does not require memory
reallocation).

Main complication coming with this is the need to keep PyBuffer object
alive for the lifetime of segyfd instance. As its .buf property points
to the certain place in memory, we also have to be sure we do not close
the buffer too early or copy the memory improperly thus changing the
.buf memory address.

buffer_guard struct was not used because it closes itself in destructor.
By using it we would have to assure correct chain of several move
constructors which carefully deal with the memory.
Using Py_buffer directly seems like a more readable option.

Py_buffer is added directly to autods as a field without creation of any
special "memory_autods" structure. Reason is an unnecessary complexity
of alternative solutions caused by bindings memory management.