Remove int-[type] specific get-set field functions
Specialized versions are not needed for the moment and we should be fine
with more general functions.

Partly reverses commits "Introduce segy_get_field_[type]" / "Introduce
segy_set_field_[type]" to make tests to use just int.
For current code it is safe because in current tests/internal functions
all values are 4 bytes or less and there are no 4-bytes unsigned
integers in current headers.

General segy_get/set_field_int integer specialization is left as is for
now as it is heavily used internally in segy.c code. We should aim to
reconsider this before 2.0 release as otherwise it would become a
permanent part of the C interface.

General int specialization is mostly used in functions dealing with
fields important for segy layout. These functions would require logic
update as those fields have overriding "extended" field. For those
general segy_get_field/segy_set_field would be used and it might present
an opportunity to get rid of "int" specialization there completely.

Why we might want to get rid of general segy_get/set_field_int integer
specialization:
- more general segy_get_field/segy_set_field function is available
- having two ways in the interface to do the same thing is confusing
- general "int" version might be dangerous in library source code as we
now want to be very careful with types of fields we read.
Now it exists purely for developer's convenience due to segyio 1.0
having only int fields.
We should aim to use segy_get_field/segy_set_field where it ok to do so.
Then we can decide whether it could be removed from the interface
completely.

Note that specialization would still be heavily used in
tests/matlab/applications, but if we manage to get rid of it in the main
code, it could become tests-only helper.