Bump minimum CMake version to 3.17
This allows us to use CMake's new way of finding and interacting with
Python. This also acts as preparatory step to move to the build process
of the Python extension from `scikit-build` to `scikit-build-core` [1]
which requires CMake >=3.15.

CMake 3.17 is chosen as minimum version because the usage of the
`WITH_SOABI` option is recommended [2] which requires CMake >=3.17 [3].

[1]: https://scikit-build-core.readthedocs.io/en/latest/
[2]: https://scikit-build-core.readthedocs.io/en/latest/guide/migration_guide.html
[3]: https://cmake.org/cmake/help/latest/module/FindPython.html#commands