Remove REQUIRE_PYTHON build option
This options did not have any observable effect because we search for
the Python packages with

```
find_package(Python COMPONENTS Interpreter REQUIRED)
```

which will issue a fatal error if it cannot find the specified PYthon
components. This means that we would never reach our own error handling
if the Python interpreter was missing. Instead, we can rely on the
`BUILD_PYTHON` build option to decide whether we want to include the
Python-specific build path or not.

(cherry picked from commit 87436acfedfd4c1ce29e38627fce2211533f2687)