Replace scikit-build by scikit-build-core
We were experiencing some issues with scikit-build on the GitHub runners
for Windows builds while also facing some deprecation warnings w.r.t.
the build system. This is solved by moving to scikit-build-core for
building the Python extension. scikit-build-core is currently more
actively developed and more modern replacement for scikit-build for
projects that do not need to manipulate the build process excessively.
The changes mostly follow the migration guide [1]. Changing to
scikit-build-core also allows us to simplify the the build process of
the Python extension.

The `Manifest.in` for building a source distribution (sdist) was removed
since it does not have any use anymore. We do not build a source
distribution at the moment. If we want to build one again, we need to
configure it inside the `pyproject.toml`.

[1]: https://scikit-build-core.readthedocs.io/en/latest/guide/migration_guide.html