Let cibuildwheel choose Linux base image
There is no obvious reason anymore to restrict the base image to
manylinux2014. Additionally, the usage of the manylinux2014 negatively
impacts the time it takes for cibuildwheel to test the wheels on
Python >=3.11 because NumPy must be build from source in this case.

(cherry picked from commit 858f4554dc05189d7432fee67fd8f35784176b73)