Check datatype match inside segy_set_field
.datatype field is kind of pointless at the moment.
- In C we can't easily force someone to set datatype the same moment
when .value is set. So nothing prevents us from setting value to i16 and
datatype to u32.
- In theory we can simply pass segy_field_value as an argument instead
of passing full segy_field_data. However it doesn't fit into the spirit
of function, as without knowing .datatype we can't interpret value on
its own and are forced to do it according to the field mapping table
instead of the actual datatype.

The best option seems to force .datatype being set by checking that it
corresponds to the datatype retrieved from field mapping.

Additionally fd doesn't have to be a pointer. We should treat that value
as any other local primitive as it is not that big in size.