Add ASCII support
Revision 1 allows data to be ASCII-encoded and segyio ability to read
ASCII data has been long requested.

Solution might not be final. It might be that we would be better off
bringing back ascii_to_ebcdic and ebcdic_to_ascii functions and doing
some conversions in segyio.cpp file instead.

Note that this commit changes default behavior for users who know their
file is ASCII and translate to it themselves.