Make offset in internal functions 0-based
0-based index is something usually expected by the developer.
Get/set field functions become more readable and less error-prone when
everything - header, mapping and offset is unified and is 0-based.

Because users are used to trace fields being 1-based and binary fields
being 3201-based we decided against changing it in public facing
functions as well. It costs us some logical repetitions, but it is
acceptable sacrifice.