Overwrite mappings by user-defined iline and xline
If user provides custom iline and xline, we assume they know what they
are doing and overwrite the mapping.
Generally there is no way we can consistently detect if values are good
or bad anymore and result of operations is unpredictable.

Only C part of the code is changed. Changing python default values and
documentation is left for later.