Add name to segy_entry_definition
At the moment we have enough information to swiftly access all field
offsets needed internally in the library by name. This is what
segy_header_mapping.name_to_offset is for.

However as we now move to user-defined xml, having names for
predefined-only entries is not enough. We do not have a structure to
store names that are not part of SEG-Y standard.

Unless we create it, we would have to disregard custom name information
on parsing. So if some file has a non-standard field in the header, user
won't know its name.

To prevent this, add NULL-terminated name information to
segy_entry_definition. It is not needed by the library itself, but would
be used in data-presenting functions.

Default names are not added to binary_map/traceheader_default_map. It is
not yet clear if it is a good idea or not.