Exchange segy_field_data datatype with entry_type
Entry_type is a more general type than our internal datatype, so it
makes more sense to store it and avoid performing internal calculations
when it becomes needed.

It also fixes the bug, that for some reason only got discovered by
valgrind check after introduction of the next commit.

When we pass fields outside of the valid range to our C functions, we
might read from the incorrect memory place as we do not check the field
range when extracting data directly from the map. We do it only inside
inner get_field/set_field functions.

By storing entry_type in segy_field_data we avoid having to deal with
this altogether.