Do not recover on bad format
It is unknown why special care was taken to deal with incorrectly set
format in binary header. Decision seems to stem from #241 but no
justification is given.
Currently we have no knowledge of files with bad format set.

We do not allow user to override other fields, like sample count and
extended text headers, so it is unclear why format is so special.

Lets assume we update the test that is supposed to check we warn user
when format is bad and that we use what we assume is the default one.
Update it by reading the trace.

def test_missing_format_ibmfloat_fallback(small):
    with segyio.open(small, mode = 'r+') as f:
        val = f.trace[0][0]
        f.bin[segyio.su.format] = 0

    with segyio.open(small) as f:
        assert int(f.format) == 1
        assert f.dtype       == np.dtype(np.float32)

        assert f.trace[0][0] == val

It fails on the old code.

As we can't even read trace correctly when we guess the format and we
didn't get lots of complains, it likely means that
files with not-set format is not a problem.

So we should be fine with dropping the support altogether.