Do not allow bad format
Lets assume we update the test that is supposed to check we warn user
when format is bad and that we use what we assume is the default one.
Update it by reading the trace.

def test_missing_format_ibmfloat_fallback(small):
    with segyio.open(small, mode = 'r+') as f:
        val = f.trace[0][0]
        f.bin[segyio.su.format] = 0

    with segyio.open(small) as f:
        assert int(f.format) == 1
        assert f.dtype       == np.dtype(np.float32)

        assert f.trace[0][0] == val

It fails on the old code.

As we can't even read trace correctly when we guess the format and we
didn't get lots of complains, it likely means that
files with not-set format simply is not a real problem.

So we should be fine with dropping the support altogether.

TODO: confirm with PO.