Ignore some compiler warnings in Python bindings
Disable warnings for methods exposed to Python arguments and potentially
missing field initializers in Python-specific data structures. Removing
the cause of the warnings is, in parts, impossible, or tedious as it
depends on the Python version the Python bindings are compiled for.

Note: The warnings became visible because we updated the build process
of the Python bindings in a previous commit. Earlier, additional
compiler flags for warnings and making warnings into errors were not
forwarded to the Python binding's build process but are forwarded now.

**Methods exposed to Python (-Wcast-function-type)**

The PyMethodDef always expects a pointer to a C implementation of type
`PyCFunction` as second argument. `PyCFunction` expects two input
arguments. If a method has only one argument, e.g., a method that only
accepts `self`, or a method with more than two arguments, e.g., a Python
method with keyword arguments, the compiler issues a warning. As far as
I understand, this cannot be resolved for functions with keyword
arguments (`METH_VARARGS | METH_KEYWORDS`, see [1]) that have three
input arguments. We can either cast the method to
`PyCFunctionWithKeywords` and get an error about too many input
arguments or we cast to PyCFunction and get an error about non-matching
function types.

**Missing field initalizers (-Wmissing-field-initializers)**

The number of fields in structs defined by the Python interface depend
on the version number. To fix this we would need, potentialy, confusing
`#if ... #else` constructions in the code.

**Guarding of pragmas**

The used pragmas are not standardised but depend on the compilers.
Therefore, we use pragmas specific to the currently detected compiler to
avoid errors about unknown pragmas

[1]: https://docs.python.org/3/c-api/structures.html
[2]: https://stackoverflow.com/a/10264563