Define D8 xml parsing function
It seems a bit complicated to parse xml in C, we are not sure of the
best approach yet.

Parsing it in python is easy though. (If we had pybind11, it probably
would be easy to translate result to C/C++ as well).

So for now parsing implementation is implemented for python binding
directly, instead of being an integral part of the C library.