Assign names to default traceheader mappings
When xml is not present, we use default maps from C, which for the
moment have their name set to NULL.

But users do not care where information came from and likely would want
to know the names as well, so they must be assigned.

---
As revision 2.1 introduced strings of unknown length and other
unknown-size features, dealing with this in C became more complicated.

Current solution where implementation happens in C++ bindings part
instead of C core is not optimal, but it works good enough in python
which is a main target user language.

There is no good vision yet what optimal solution would be.