Set endianness before extended text headers parse
Stanzas were gathered incorrectly for little endian files as endianness
was still assumed to be big-endian.

There is a bit of a design problem.

Operations flow that we need is
- figure out endianness
- read declared number of extended text headers
- if it is -1, calculate number of stanzas
- proceed with finding file metadata and stanzas

This breaks our attempt to combine all metadata retrievals in one C
function, as we don't read stanzas in C.

Until it is decided upon better design, explicitly set endianness in
segyopen before all other operations.