Learn to read files with trace header extensions
trace-header-extensions.sgy file is created manually, and will be
updated manually when more features are needed. In the future it should
be possible to recreate it with a script.
In this commit it is enough that segyio can read it.

Binary applications are not updated and still work on revision 1 only.