Calculate delay time in segy.c
Both DelayRecordingTime and ScalarTraceHeader are used in open.
With custom mapping those two can have any type. In other internal
functions used in open we restrict the type in segy.c so that we have a
correct state.

So it makes sense to move calculations to C, so that both tightly
coupled dt and delay were read and processes nearby.