Add Field wrapper to support access by names
New class had to be added to keep performance of the old Field class the
same.

Introduction of __getattr__ and __setattr__ directly to Field causes
decrease in performance compared with old code.

Nevertheless, we still want to give user the chance to retrieve data via
names and out of all approaches discussed in the team __getattr__ /
__setattr__ appeared to be the most pythonic and consistent.

So new class HeaderFieldAccessor was introduced on top of the Field one.

New naming scheme is chosen to
- support level-like access to the data
- distinguish this trace-number based hierarchy of field values from
trace-independent hierarchy of layout information
- alleviate confusion of names like Header/Field which actually are
lists/sequences/tables of headers/fields.